agent/multiagent/pyproject.toml
[project]
name = "sales-multiagent"
version = "0.1.0"
description = "Multi-agent orchestrator for Enterprise Sales Analytics with Genie subagent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "databricks-openai>=0.11.0",
    "databricks-agents",
    "mlflow>=3.9.0",
    "openai-agents>=0.4.1",
    "python-dotenv",
    "pyyaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_server", "scripts"]

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "pytest>=7.0.0",
]

[project.scripts]
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
